Move state helpers to shared file
