--- packages/react-urql/src/hooks/constants.ts
@@ -1,8 +1,0 @@
-export const initialState = {
-  fetching: false,
-  stale: false,
-  error: undefined,
-  data: undefined,
-  extensions: undefined,
-  operation: undefined,
-};
